// src/app/(system)/(service)/adminUser/types.ts
export type DepartmentListVos = {
  id: string;
  name: string;
};

export type CommunityInner = {
  id: string;
  communityName: string;
};

/**
 * @description View Type
 */
export type AdminUserVo = {
  id: string; // id
  userId: string; // userId
  firstName: string; // firstName
  lastName: string; // lastName
  email: string; // email
  phone: string; // phone
  title: string; // title
  terminationDate: string; // terminationDate
  status: keyof typeof ADMIN_USER_STATUS; // status
  statusName: string; // statusName
  community: CommunityInner[]; // community
  departmentListVos: DepartmentListVos[];
  displayNames: string;
  createdAt: string; // createdAt
  createdBy: string; // createdBy
  updatedAt: string; // updatedAt
  updatedBy: string; // updatedBy
  isDeleted: boolean; // isDeleted
  communityId: string; // communityId
};